requirements.txt
# ============================================================================
# AI Investment Advisor - Python Dependencies
# ============================================================================

# ----------------------------------------------------------------------------
# Core Framework
# ----------------------------------------------------------------------------
fastapi==0.104.1              # Modern web framework for APIs
uvicorn[standard]==0.24.0     # ASGI server with WebSockets support
pydantic==2.5.0               # Data validation using Python type hints
python-dotenv==1.0.0          # Environment variable management

# ----------------------------------------------------------------------------
# Database
# ----------------------------------------------------------------------------
psycopg2-binary==2.9.9        # PostgreSQL adapter for Python
# Note: Use psycopg2-binary for development/testing
# For production, use psycopg2 (requires compilation)

# ----------------------------------------------------------------------------
# LLM Providers
# ----------------------------------------------------------------------------
requests==2.31.0              # HTTP library for API calls
# Required for both Ollama and Groq providers

# ----------------------------------------------------------------------------
# Sentiment Analysis
# ----------------------------------------------------------------------------
# Option 1: VADER (Fast, Rule-based) - RECOMMENDED FOR TESTING
vaderSentiment==3.3.2

# Option 2: FinBERT (Accurate, ML-based) - OPTIONAL
# Uncomment if you want to use FinBERT for sentiment analysis
# transformers==4.35.0        # Hugging Face transformers
# torch==2.1.0                # PyTorch (required for transformers)
# Note: torch is large (~2GB), only install if using FinBERT

# ----------------------------------------------------------------------------
# Development & Testing
# ----------------------------------------------------------------------------
pytest==7.4.3                 # Testing framework
pytest-asyncio==0.21.1        # Async support for pytest
pytest-cov==4.1.0             # Coverage plugin for pytest

# Code Quality
black==23.11.0                # Code formatter
flake8==6.1.0                 # Linter
isort==5.12.0                 # Import sorter
mypy==1.7.1                   # Type checker

# ----------------------------------------------------------------------------
# Optional: Data Analysis & Visualization
# ----------------------------------------------------------------------------
# Uncomment if you want to add data analysis capabilities
# pandas==2.1.3               # Data manipulation
# numpy==1.26.2               # Numerical computing
# matplotlib==3.8.2           # Plotting
# seaborn==0.13.0             # Statistical visualization

# ----------------------------------------------------------------------------
# Optional: Advanced Features
# ----------------------------------------------------------------------------
# Uncomment if you need these features

# Redis for caching
# redis==5.0.1
# hiredis==2.2.3              # Fast Redis parser

# Celery for distributed task queue
# celery==5.3.4
# celery[redis]==5.3.4

# Advanced monitoring
# prometheus-client==0.19.0
# opentelemetry-api==1.21.0

# Rate limiting
# slowapi==0.1.9

# ============================================================================
# Installation Instructions
# ============================================================================
#
# Basic Installation (Recommended for testing):
#   pip install -r requirements.txt
#
# With FinBERT (adds ~3GB):
#   pip install -r requirements.txt
#   pip install transformers torch
#
# Development Installation:
#   pip install -r requirements.txt
#   # All dev dependencies already included
#
# Production Installation:
#   pip install -r requirements.txt
#   # Consider using psycopg2 instead of psycopg2-binary
#   pip uninstall psycopg2-binary
#   pip install psycopg2
#
# ============================================================================
# Python Version
# ============================================================================
# Requires: Python 3.8 or higher
# Tested on: Python 3.9, 3.10, 3.11
# Recommended: Python 3.11 (best performance)
#
# ============================================================================
# Platform-Specific Notes
# ============================================================================
#
# Windows:
#   - psycopg2-binary works out of the box
#   - If using torch: `pip install torch --index-url https://download.pytorch.org/whl/cpu`
#
# macOS (M1/M2):
#   - psycopg2-binary works on Apple Silicon
#   - torch: Use conda for better performance
#     `conda install pytorch torchvision -c pytorch`
#
# Linux:
#   - All packages work out of the box
#   - For production, compile psycopg2 from source
#
# ============================================================================
# Docker
# ============================================================================
# If using Docker, see Dockerfile for optimized installation
#
# ============================================================================
# Troubleshooting
# ============================================================================
#
# Issue: psycopg2 installation fails
# Solution: Use psycopg2-binary OR install PostgreSQL dev headers
#   Ubuntu/Debian: sudo apt-get install libpq-dev
#   CentOS/RHEL: sudo yum install postgresql-devel
#   macOS: brew install postgresql
#
# Issue: torch too large
# Solution: Install CPU-only version
#   pip install torch --index-url https://download.pytorch.org/whl/cpu
#
# Issue: ImportError for a package
# Solution: Reinstall in a fresh virtual environment
#   python -m venv venv
#   source venv/bin/activate  # On Windows: venv\Scripts\activate
#   pip install --upgrade pip
#   pip install -r requirements.txt
#
# ============================================================================